- Fix Zip Slip vulnerability - Upgrade FontAwesome version to 5.0.13 - Upgrade to gradle 4.8

+++ slideshowfx-utils/src/main/java/com/twasyl/slideshowfx/utils/ZipUtils.java
@@ -20,15 +20,16 @@ public class ZipUtils {
     private static final Logger LOGGER = Logger.getLogger(ZipUtils.class.getName());
 
     /**
-     * Unzpi the given archive into the provided destination. If the destination does not exist it is created.
-     * @param archive The archive file to unzip.
+     * Unzip the given archive into the provided destination. If the destination does not exist it is created.
+     *
+     * @param archive     The archive file to unzip.
      * @param destination The destination directory where the archive will be unzipped.
-     * @throws IOException If the archive file does not exist.
-     * @throws java.lang.NullPointerException If the archive file or the destination is null.
+     * @throws IOException          If the archive file does not exist.
+     * @throws NullPointerException If the archive file or the destination is null.
      */
     public static void unzip(File archive, File destination) throws IOException, IllegalArgumentException {
-        if(archive == null) throw new NullPointerException("The ZIP file can not be null");
-        if(!archive.exists()) throw new FileNotFoundException("The ZIP file does not exist");
+        if (archive == null) throw new NullPointerException("The ZIP file can not be null");
+        if (!archive.exists()) throw new FileNotFoundException("The ZIP file does not exist");
 
         // Unzip
         LOGGER.fine("Extracting file " + archive.toURI().toASCIIString());
@@ -38,53 +39,58 @@ public class ZipUtils {
     }
 
     /**
-     * Unzpi the given archive into the provided destination. If the destination does not exist it is created.
-     * @param archive The archive file to unzip.
+     * Unzip the given archive into the provided destination. If the destination does not exist it is created.
+     *
+     * @param archive     The archive file to unzip.
      * @param destination The destination directory where the archive will be unzipped.
-     * @throws IOException If the archive file does not exist.
-     * @throws java.lang.NullPointerException If the archive file or the destination is null.
+     * @throws IOException          If the archive file does not exist.
+     * @throws NullPointerException If the archive file or the destination is null.
      */
     public static void unzip(InputStream archive, File destination) throws IOException, IllegalArgumentException {
-        if(archive == null) throw new NullPointerException("The ZIP file can not be null");
-        if(destination == null) throw new NullPointerException("The destination can not be null");
+        if (archive == null) throw new NullPointerException("The ZIP file can not be null");
+        if (destination == null) throw new NullPointerException("The destination can not be null");
 
-        if(!destination.exists()) {
-            if(!destination.mkdirs()) {
+        if (!destination.exists()) {
+            if (!destination.mkdirs()) {
                 throw new IOException("Can not create destination folder");
             }
         }
 
-        ZipInputStream zipReader = new ZipInputStream(archive);
+        final ZipInputStream zipReader = new ZipInputStream(archive);
         ZipEntry zipEntry;
         File extractedFile;
 
-        while((zipEntry = zipReader.getNextEntry()) != null) {
+        while ((zipEntry = zipReader.getNextEntry()) != null) {
             extractedFile = new File(destination, zipEntry.getName());
 
-            LOGGER.fine("Extracting file: " + extractedFile.getAbsolutePath());
 
-            if(zipEntry.isDirectory()) {
-                if(!extractedFile.exists() && !extractedFile.mkdirs()) {
-                    throw new IOException("Can not create folder");
-                }
-            }
-            else {
-                // Ensure to create the parents directories
-                if(!extractedFile.getParentFile().exists() && !extractedFile.getParentFile().mkdirs()) {
-                    throw new IOException("Can not create the parent folder");
-                }
+            if (isFileInParent(extractedFile, destination)) {
+                LOGGER.fine("Extracting file: " + extractedFile.getAbsolutePath());
 
-                int length;
-                byte[] buffer = new byte[1024];
-
-                try(final FileOutputStream extractedFileOutputStream = new FileOutputStream(extractedFile)) {
-
-                    while ((length = zipReader.read(buffer)) > 0) {
-                        extractedFileOutputStream.write(buffer, 0, length);
+                if (zipEntry.isDirectory()) {
+                    if (!extractedFile.exists() && !extractedFile.mkdirs()) {
+                        throw new IOException("Can not create folder");
+                    }
+                } else {
+                    // Ensure to create the parents directories
+                    if (!extractedFile.getParentFile().exists() && !extractedFile.getParentFile().mkdirs()) {
+                        throw new IOException("Can not create the parent folder");
                     }
 
-                    extractedFileOutputStream.flush();
+                    int length;
+                    byte[] buffer = new byte[1024];
+
+                    try (final FileOutputStream extractedFileOutputStream = new FileOutputStream(extractedFile)) {
+
+                        while ((length = zipReader.read(buffer)) > 0) {
+                            extractedFileOutputStream.write(buffer, 0, length);
+                        }
+
+                        extractedFileOutputStream.flush();
+                    }
                 }
+            } else {
+                LOGGER.severe("Skipping unzipping entry " + zipEntry.getName() + " as it isn't in the destination");
             }
         }
 
@@ -96,15 +102,16 @@ public class ZipUtils {
 
     /**
      * Compress the given fileToZip into the given destination. This method manages if the fileToZip is a folder or a simple file.
-     * @param fileToZip The content to compress.
+     *
+     * @param fileToZip   The content to compress.
      * @param destination The destination into the content will be compressed.
-     * @throws java.lang.NullPointerException If the fileToZip or the destination is null.
-     * @throws java.io.FileNotFoundException If the fileToZip does not exist.
+     * @throws NullPointerException  If the fileToZip or the destination is null.
+     * @throws FileNotFoundException If the fileToZip does not exist.
      */
     public static void zip(File fileToZip, File destination) throws IOException {
-        if(fileToZip == null) throw new NullPointerException("The file to zip can not be null");
-        if(!fileToZip.exists()) throw new FileNotFoundException("The file to zip does not exist");
-        if(destination == null) throw new NullPointerException("The destination can not be null");
+        if (fileToZip == null) throw new NullPointerException("The file to zip can not be null");
+        if (!fileToZip.exists()) throw new FileNotFoundException("The file to zip does not exist");
+        if (destination == null) throw new NullPointerException("The destination can not be null");
 
         final ListFilesFileVisitor visitor = new ListFilesFileVisitor();
         Files.walkFileTree(fileToZip.toPath(), visitor);
@@ -114,24 +121,24 @@ public class ZipUtils {
         ZipOutputStream zipOutput = new ZipOutputStream(new FileOutputStream(destination));
         ZipEntry entry;
         String entryName;
-        byte[] buffer  = new byte[1024];
+        byte[] buffer = new byte[1024];
         int length;
 
         String prefixToDelete;
-        if(fileToZip.isDirectory()) {
+        if (fileToZip.isDirectory()) {
             prefixToDelete = fileToZip.getAbsolutePath() + File.separator;
         } else {
             prefixToDelete = "";
         }
 
-        for(File file : filesToZip) {
+        for (File file : filesToZip) {
             LOGGER.fine("Compressing file: " + file.getAbsolutePath());
 
             entryName = file.getAbsolutePath().substring(prefixToDelete.length(), file.getAbsolutePath().length());
             entryName = entryName.replaceAll("\\\\", "/");
             LOGGER.finest("Entry name: " + entryName);
 
-            if(file.isDirectory()) {
+            if (file.isDirectory()) {
                 entry = new ZipEntry(entryName + "/");
                 zipOutput.putNextEntry(entry);
             } else {
@@ -141,7 +148,7 @@ public class ZipUtils {
                 try {
                     fileInput = new FileInputStream(file);
 
-                    while((length = fileInput.read(buffer)) > 0) {
+                    while ((length = fileInput.read(buffer)) > 0) {
                         zipOutput.write(buffer, 0, length);
                     }
                 } finally {
@@ -155,5 +162,17 @@ public class ZipUtils {
         zipOutput.close();
 
         LOGGER.fine("File compressed");
+    }
+
+    /**
+     * Tests if the given file is contained within a given parent folder.
+     *
+     * @param file   The file to test the presence in the parent.
+     * @param parent The parent that should contain the file.
+     * @return {@code true} if the file is within the parent, {@code false} otherwise.
+     * @throws IOException If something went wrong during the check.
+     */
+    protected static boolean isFileInParent(final File file, final File parent) throws IOException {
+        return file.getCanonicalPath().startsWith(parent.getCanonicalPath());
     }
 }
